move css value list arrow handling to ds
