+++ apps/builder/app/builder/features/style-panel/sections/backgrounds/backgrounds.tsx
@@ -3,9 +3,8 @@ import type { RenderCategoryProps } from "../../style-sections";
 import { styleConfigByName } from "../../shared/configs";
 import { FloatingPanel } from "~/builder/shared/floating-panel";
 import {
-  ArrowFocus,
-  Box,
   CssValueListItem,
+  CssValueListItemsArrowFocus,
   Flex,
   Grid,
   Label,
@@ -43,7 +42,7 @@ import {
 import { BackgroundContent } from "./background-content";
 import { getLayerName, LayerThumbnail } from "./background-thumbnail";
 import { useSortable } from "./use-sortable";
-import { useMemo, type ReactNode } from "react";
+import { useMemo } from "react";
 import type { RgbValue, StyleProperty } from "@webstudio-is/css-data";
 import {
   CollapsibleSectionBase,
@@ -51,13 +50,9 @@ import {
 } from "~/builder/shared/collapsible-section";
 import { getDots } from "../../shared/collapsible-section";
 
-const LIST_ITEM_ATTRIBUTE = "data-list-item";
-
-const listItemAttributes = { [LIST_ITEM_ATTRIBUTE]: true };
-
 const Layer = (props: {
   id: string;
-  tabIndex: -1 | 0;
+  index: number;
   isHighlighted: boolean;
   layerStyle: StyleInfo;
   setProperty: SetBackgroundProperty;
@@ -105,7 +100,7 @@ const Layer = (props: {
       <CssValueListItem
         active={props.isHighlighted}
         data-id={props.id}
-        tabIndex={props.tabIndex}
+        index={props.index}
         label={
           <Label truncate onReset={props.deleteLayer}>
             {getLayerName(props.layerStyle, assets)}
@@ -132,7 +127,6 @@ const Layer = (props: {
             />
           </>
         }
-        {...listItemAttributes}
       />
     </FloatingPanel>
   );
@@ -202,28 +196,6 @@ const BackgroundsCollapsibleSection = ({
   );
 };
 
-const ListItemsFocusWrap = (props: { children: ReactNode }) => {
-  return (
-    <ArrowFocus
-      render={({ handleKeyDown }) => (
-        <Box
-          css={{ display: "contents" }}
-          onKeyDown={(event) => {
-            if (event.key === "ArrowUp" || event.key === "ArrowDown") {
-              handleKeyDown(event, {
-                accept: (element) =>
-                  element.getAttribute(LIST_ITEM_ATTRIBUTE) === "true",
-              });
-            }
-          }}
-        >
-          {props.children}
-        </Box>
-      )}
-    />
-  );
-};
-
 export const BackgroundsSection = (props: RenderCategoryProps) => {
   const { setProperty, deleteProperty, currentStyle, createBatchUpdate } =
     props;
@@ -256,7 +228,7 @@ export const BackgroundsSection = (props: RenderCategoryProps) => {
       category={props.category}
     >
       <Flex gap={1} direction="column">
-        <ListItemsFocusWrap>
+        <CssValueListItemsArrowFocus>
           <Flex
             gap={1}
             direction="column"
@@ -270,7 +242,7 @@ export const BackgroundsSection = (props: RenderCategoryProps) => {
             {layers.map((layer, index) => (
               <Layer
                 id={layer.id}
-                tabIndex={index === 0 ? 0 : -1}
+                index={index}
                 key={layer.id}
                 isHighlighted={dragItemId === layer.id}
                 layerStyle={getLayerBackgroundStyleInfo(
@@ -299,7 +271,7 @@ export const BackgroundsSection = (props: RenderCategoryProps) => {
 
             {placementIndicator}
           </Flex>
-        </ListItemsFocusWrap>
+        </CssValueListItemsArrowFocus>
 
         <Grid
           css={{
