+++ apps/builder/app/builder/features/style-panel/sections/box-shadows/box-shadow-list.tsx
@@ -1,5 +1,5 @@
 import type { LayersValue } from "@webstudio-is/css-data";
-import { Flex } from "@webstudio-is/design-system";
+import { CssValueListItemsArrowFocus, Flex } from "@webstudio-is/design-system";
 import type { RenderCategoryProps } from "../../style-sections";
 import { Layer } from "./box-shadow-layer";
 import { deleteLayer, hideLayer, updateBoxShadowLayer } from "./utils";
@@ -25,23 +25,25 @@ export const BoxShadowLayersList = ({
   };
 
   return (
-    <Flex direction="column" gap={2}>
-      {layers.value.map((layer, index) => {
-        if (layer.type === "tuple") {
-          return (
-            <Layer
-              key={index}
-              index={index}
-              layer={layer}
-              onLayerHide={handleHideLayer}
-              onDeleteLayer={handleDeleteLayer}
-              createBatchUpdate={createBatchUpdate}
-              onEditLayer={onEditLayer}
-            />
-          );
-        }
-        return null;
-      })}
-    </Flex>
+    <CssValueListItemsArrowFocus>
+      <Flex direction="column" gap={2}>
+        {layers.value.map((layer, index) => {
+          if (layer.type === "tuple") {
+            return (
+              <Layer
+                key={index}
+                index={index}
+                layer={layer}
+                onLayerHide={handleHideLayer}
+                onDeleteLayer={handleDeleteLayer}
+                createBatchUpdate={createBatchUpdate}
+                onEditLayer={onEditLayer}
+              />
+            );
+          }
+          return null;
+        })}
+      </Flex>
+    </CssValueListItemsArrowFocus>
   );
 };
